deleteItem
