+++ pages/deleteItem/index.html
@@ -12,13 +12,13 @@
     <link href="/index.css" rel="stylesheet" />
     <script src="/libraries/jquery/jquery-3.5.1.min.js"></script>
     <script src="/libraries/bootstrap/js/bootstrap.min.js"></script>
-    <title>Employee Login</title>
+    <title>Item Delete</title>
   </head>
   <body class="indexBackground">
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <nav-bar></nav-bar>
-    <add-item></add-item>
+    <delete-item></delete-item>
   </body>
-  <script src="../modules/navbar.js" type="module"></script>
-  <script src="../modules/deleteItem.js" type="module"></script>
+  <script src="/modules/navbar.js" type="module"></script>
+  <script src="/pages/deleteItem/deleteItem.js" type="module"></script>
 </html>
